Add hook for letting users choose what features to use
With this change we're opening up for the possibility to require just the parts
of rails-mailpack that you actually want to use in your app.

To begin with, we're supporting

    require "rails/mailpack/all"

that imports everything.

Already now the user could do for example

    # Gemfile
    gem "rails-mailpack", :require => false

    # config/initializers/mailpack.rb
    require "markerb"

to require just the markerb stuff. They'd probably be better off just requiring
the markerb gem directly, though. A better example might be when you don't want
for example markerb, you could do

    # Gemfile
    gem "rails-mailpack", :require => false

    # config/initializers/mailpack.rb
    require "html_links_to_text_interceptor"
    require "premailer/rails"

which would prevent markerb from being loaded.